% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations for sample selection}

\item{n}{sample size}

\item{p}{probability of success}
}
\value{
a barplot representing the distribution
}
\description{
Title
}
\examples{
mybin(iter = 10000, n = 30, p = 0.6)

}
