% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{data to be used}

\item{fun}{statistic used to create bootstrap intervals}

\item{alpha}{1 - alpha \% confidence interval}

\item{cx}{text font size}

\item{...}{any other parameters}
}
\value{
a histogram of the data with labled mean and bootstrap interval
}
\description{
Title
}
\examples{
myboot2(x = fire$DAMAGE, alpha = 0.05)
 myboot2(iter = 2000, x = fire$DAMAGE)

}
