% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats on the flight}

\item{gamma}{probability a plane is indeed overbooked}

\item{p}{probability of a show}
}
\value{
2 plots (discrete and continuous) and a list of variable values
}
\description{
Title
}
