% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Title}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iteratoins}

\item{p}{probability of success}

\item{...}{...}
}
\value{
a histogram of the mean of the binomial distribution
}
\description{
Title
}
