% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{sample mean}

\item{sigma}{sample standard deviation}

\item{a}{upper limit}
}
\value{
a plot of a normal distribution with the area from negative infinity to a colored
}
\description{
Title
}
\examples{
myncurve(mu = 10, sigma = 5, a = 6)

}
